requirements.txt
# VoqTask - Voice To-Do Task Manager (Streamlit)
streamlit>=1.28.0
openai-whisper>=20231117
soundfile>=0.12.1
numpy>=1.24.0
# Whisper depends on torch; install separately for GPU: pip install torch
# Optional: lighter STT fallback when offline / no Whisper
SpeechRecognition>=3.10.0
pydub>=0.25.1
# Storage & sync (optional)
supabase>=2.0.0
python-dotenv>=1.0.0
# Optional: TTS for "Read aloud" (SDG 10)
pyttsx3>=2.90
